+++ Automate/docs/release-notes.md
@@ -4,7 +4,9 @@
 ## Upcoming release
 * Automate now rebuilds machine data at the start of each day, to fix edge cases like building upgrades.
 * Added `GetFilled()` and `GetCapacity()` methods to `IContainer` interface. This affects mods which add custom container types.
-* The `automate summary` command now shows machine states and chest fullness.
+* Improved `automate summary` console command:
+  * it now shows machine states and chest fullness;
+  * it now shows a warning if it detects a common cause of lag.
 * Refactored console command handling.
 
 ## 1.23.2